refactor(BridgeGame) 파일, 클래스명 변경
Changed:
- BridgeService -> BridgeGame
